aula-pi-main/backend/requirements.txt
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
blinker==1.9.0
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
click==8.1.7
Flask==2.3.3
Flask-Cors==5.0.0
idna==3.10
importlib_metadata==8.5.0
iniherit==0.3.9
itsdangerous==2.1.2
Jinja2==3.1.4
MarkupSafe==2.1.5
marshmallow==3.23.1
mysql-connector-python==9.1.0
packaging==24.2
pycparser==2.22
PyJWT==2.10.0
python-dotenv==1.0.1
pytz==2024.2
requests==2.32.3
resend==2.4.0
six==1.16.0
sqlparse==0.5.2
tabulate==0.8.2
text-unidecode==1.3
tomli==2.0.1
typing_extensions==4.12.2
urllib3==1.26.20
Werkzeug==3.0.4
yoyo-migrations==9.0.0
zipp==3.21.0
